refactor(#1953): slice P1 — relocate the narrowed-LLM engine to TaskExecutionHost
P1 of the planner→Task subsumption (spec F): a byte-identical MOVE of the
narrowed-LLM execution engine out of planner.py into its own Task-execution-engine
home, so P2 can generalize it to run a Task (not only a plan step).

- `_PlanStepHost` (+ its tool-family-gate constants) relocates char-identical into
  `src/reyn/runtime/task_execution.py` as `TaskExecutionHost`. The class body is
  unchanged (only the class name + a doc-comment reference); the #1984 qualified-name
  family-gates, the force-close / retry interface, and the dispatch passthroughs
  move verbatim.
- The planner keeps driving it under its historical name via a **coexist adapter**
  (`from reyn.runtime.task_execution import TaskExecutionHost` + `_PlanStepHost =
  TaskExecutionHost`), so plan behavior is byte-identical — the per-step driver
  (sub-loop build / retry / force-close re-entry) still lives in planner.py and
  moves in P2.

This is a relocation, not a rewrite: the gate is byte-identical per-unit behavior,
proven by the 140 existing planner tests passing unchanged.

Tests: the P1 relocation contract (the engine lives in task_execution as
TaskExecutionHost; the planner drives it via the coexist alias; the relocated
engine's narrowing still works). Full suite 8247 passed (only the 2 known
not-my-diff env failures); planner suite (140) byte-identical green.
